# server/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "CDC-CBIT API",
    "description": "The following is the list of all the endpoints available for different end-users."
  },
  "host": "localhost:3080",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "any description..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/snaps/": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/snaps/{snapId}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "snapId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "delete": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "snapId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/snaps/placement_snaps/{placementId}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "placementId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/snaps/{placementId}": {
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "placementId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "snap_name": {
                  "type": "string",
                  "example": "any"
                },
                "type_of_data": {
                  "type": "string",
                  "example": "any"
                },
                "placement_batch": {
                  "type": "string",
                  "example": "any"
                },
                "fields": {
                  "type": "string",
                  "example": "any"
                },
                "extra_fields": {
                  "type": "string",
                  "example": "any"
                },
                "last_date": {
                  "type": "string",
                  "example": "any"
                },
                "isFirst": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/snaps/add_snap/{snapId}": {
      "patch": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "snapId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/snaps/add_data/{snapId}": {
      "patch": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "snapId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/snaps/remove_data/{snapId}": {
      "patch": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "snapId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/snaps/update_data/{snapId}": {
      "patch": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "snapId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/snaps/snap_data/{plcId}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "plcId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/snaps/download/{snapId}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "snapId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
...
}